+++ openstack/elb/v3/flavors/List.go
@@ -1,7 +1,7 @@
 package flavors
 
 import (
-	golangsdk "github.com/opentelekomcloud/gophertelekomcloud"
+	"github.com/opentelekomcloud/gophertelekomcloud"
 	"github.com/opentelekomcloud/gophertelekomcloud/pagination"
 )
 
@@ -46,10 +46,4 @@ func List(client *golangsdk.ServiceClient, opts ListOptsBuilder) pagination.Page
 	return pagination.NewPager(client, url, func(r pagination.PageResult) pagination.Page {
 		return FlavorPage{PageWithInfo: pagination.NewPageWithInfo(r)}
 	})
-}
-
-// Get returns additional information about a Flavor, given its ID.
-func Get(client *golangsdk.ServiceClient, flavorID string) (r GetResult) {
-	_, r.Err = client.Get(client.ServiceURL("flavors", flavorID), &r.Body, nil)
-	return
 }
