+++ openstack/elb/v3/policies/requests.go
@@ -1,9 +1,7 @@
 package policies
 
 import (
-	golangsdk "github.com/opentelekomcloud/gophertelekomcloud"
 	"github.com/opentelekomcloud/gophertelekomcloud/openstack/elb/v3/rules"
-	"github.com/opentelekomcloud/gophertelekomcloud/pagination"
 )
 
 type Action string
@@ -93,171 +91,4 @@ type RedirectUrlOptions struct {
 	//
 	// The value can be 301, 302, 303, 307, or 308.
 	StatusCode string `json:"status_code" required:"true"`
-}
-
-type CreateOpts struct {
-	// Specifies where requests will be forwarded. The value can be one of the following:
-	//
-	//    REDIRECT_TO_POOL: Requests will be forwarded to another backend server group.
-	//    REDIRECT_TO_LISTENER: Requests will be redirected to an HTTPS listener.
-	Action Action `json:"action" required:"true"`
-
-	// Specifies the conditions contained in a forwarding rule.
-	// This parameter will take effect when enhance_l7policy_enable is set to true.
-	// If conditions is specified, key and value will not take effect,
-	// and the value of this parameter will contain all conditions configured for the forwarding rule.
-	// The keys in the list must be the same, whereas each value must be unique.
-	Conditions []rules.Condition `json:"conditions,omitempty"`
-
-	// Provides supplementary information about the forwarding policy.
-	Description string `json:"description,omitempty"`
-
-	// Specifies the configuration of the page that will be returned.
-	// This parameter will take effect when
-	FixedResponseConfig *FixedResponseOptions `json:"fixed_response_config,omitempty"`
-
-	// Specifies the ID of the listener to which the forwarding policy is added.
-	ListenerID string `json:"listener_id" required:"true"`
-
-	// Specifies the forwarding policy name.
-	Name string `json:"name,omitempty"`
-
-	// Specifies the forwarding policy priority. The value cannot be updated.
-	Position int `json:"position,omitempty"`
-
-	// Specifies the forwarding policy priority. The value cannot be updated.
-	Priority int `json:"priority,omitempty"`
-
-	// Specifies the ID of the project where the forwarding policy is used.
-	ProjectID string `json:"project_id,omitempty"`
-
-	// Specifies the ID of the listener to which requests are redirected.
-	// This parameter is mandatory when action is set to REDIRECT_TO_LISTENER.
-	RedirectListenerID string `json:"redirect_listener_id,omitempty"`
-
-	// Specifies the ID of the backend server group that requests are forwarded to.
-	//
-	// This parameter is valid and mandatory only when action is set to REDIRECT_TO_POOL.
-	// The specified backend server group cannot be the default one associated with the listener,
-	// or any backend server group associated with the forwarding policies of other listeners.
-	RedirectPoolID string `json:"redirect_pool_id,omitempty"`
-
-	// Specifies the URL to which requests are forwarded.
-	//
-	// Format: protocol://host:port/path?query
-	RedirectUrl string `json:"redirect_url,omitempty"`
-
-	// Specifies the URL to which requests are forwarded.
-	// This parameter is mandatory when action is set to REDIRECT_TO_URL.
-	// It cannot be specified if the value of action is not REDIRECT_TO_URL.
-	RedirectUrlConfig *RedirectUrlOptions `json:"redirect_url_config,omitempty"`
-
-	// Lists the forwarding rules in the forwarding policy.
-	// The list can contain a maximum of 10 forwarding rules (if conditions is specified, a condition is considered as a rule).
-	Rules []Rule `json:"rules,omitempty"`
-
-	// Specifies the configuration of the backend server group that the requests are forwarded to. This parameter is valid only when action is set to REDIRECT_TO_POOL.
-	RedirectPoolsConfig []RedirectPoolOptions `json:"redirect_pools_config,omitempty"`
-}
-
-type CreateOptsBuilder interface {
-	ToPolicyCreateMap() (map[string]interface{}, error)
-}
-
-func (opts CreateOpts) ToPolicyCreateMap() (map[string]interface{}, error) {
-	return golangsdk.BuildRequestBody(opts, "l7policy")
-}
-
-func Create(client *golangsdk.ServiceClient, opts CreateOptsBuilder) (r CreateResult) {
-	b, err := opts.ToPolicyCreateMap()
-	if err != nil {
-		r.Err = err
-		return
-	}
-	_, r.Err = client.Post(client.ServiceURL("l7policies"), b, &r.Body, nil)
-	return
-}
-
-type ListOptsBuilder interface {
-	ToPolicyListQuery() (string, error)
-}
-
-type ListOpts struct {
-	Marker             string   `q:"marker"`
-	Limit              int      `q:"limit"`
-	PageReverse        bool     `q:"page_reverse"`
-	ID                 []string `q:"id"`
-	Name               []string `q:"name"`
-	Description        []string `q:"description"`
-	ListenerID         []string `q:"listener_id"`
-	Action             []string `q:"action"`
-	RedirectPoolID     []string `q:"redirect_pool_id"`
-	RedirectListenerID []string `q:"redirect_listener_id"`
-	ProvisioningStatus []string `q:"provisioning_status"`
-	DisplayAllRules    bool     `q:"display_all_rules"`
-}
-
-func (opts ListOpts) ToPolicyListQuery() (string, error) {
-	q, err := golangsdk.BuildQueryString(opts)
-	if err != nil {
-		return "", err
-	}
-	return q.String(), nil
-}
-
-func List(client *golangsdk.ServiceClient, opts ListOptsBuilder) pagination.Pager {
-	url := client.ServiceURL("l7policies")
-	if opts != nil {
-		q, err := opts.ToPolicyListQuery()
-		if err != nil {
-			return pagination.Pager{Err: err}
-		}
-		url += q
-	}
-
-	return pagination.NewPager(client, url, func(r pagination.PageResult) pagination.Page {
-		return PolicyPage{PageWithInfo: pagination.NewPageWithInfo(r)}
-	})
-}
-
-func Get(client *golangsdk.ServiceClient, id string) (r GetResult) {
-	_, r.Err = client.Get(client.ServiceURL("l7policies", id), &r.Body, nil)
-	return
-}
-
-type UpdateOpts struct {
-	Name                *string               `json:"name,omitempty"`
-	Description         *string               `json:"description,omitempty"`
-	RedirectListenerID  string                `json:"redirect_listener_id,omitempty"`
-	RedirectPoolID      string                `json:"redirect_pool_id,omitempty"`
-	Rules               []Rule                `json:"rules,omitempty"`
-	RedirectUrlConfig   *RedirectUrlOptions   `json:"redirect_url_config,omitempty"`
-	FixedResponseConfig *FixedResponseOptions `json:"fixed_response_config,omitempty"`
-	Priority            int                   `json:"priority,omitempty"`
-	RedirectPoolsConfig []RedirectPoolOptions `json:"redirect_pools_config,omitempty"`
-}
-
-type UpdateOptsBuilder interface {
-	ToPolicyUpdateMap() (map[string]interface{}, error)
-}
-
-func (opts UpdateOpts) ToPolicyUpdateMap() (map[string]interface{}, error) {
-	return golangsdk.BuildRequestBody(opts, "l7policy")
-}
-
-func Update(client *golangsdk.ServiceClient, id string, opts UpdateOptsBuilder) (r UpdateResult) {
-	b, err := opts.ToPolicyUpdateMap()
-	if err != nil {
-		r.Err = err
-		return
-	}
-	_, r.Err = client.Put(client.ServiceURL("l7policies", id), b, &r.Body, &golangsdk.RequestOpts{
-		OkCodes: []int{200, 201},
-	})
-	return
-}
-
-func Delete(client *golangsdk.ServiceClient, id string) (r DeleteResult) {
-	_, r.Err = client.Delete(client.ServiceURL("l7policies", id), nil)
-	return
 }
